Comment comilation and syntax errors

+++ src/main/java/cs203/g1t7/users/UserService.java
@@ -4,13 +4,13 @@ import java.util.List;
 
 public interface UserService {
     List<User> listUsers();
-    User getUser(Long id);
+    User getUser(Integer id);
     User addUser(User user);
-    User updateUser(Long id, User user);
+    User updateUser(Integer id, User user);
 
     /**
      * Change method's signature: do not return a value for delete operation
      * @param id
      */
-    void deleteUser(Long id);
+    void deleteUser(Integer id);
 }
